refactor: correct mismatches between component name and file name
